refactor: 추후 사용자 입력값을 api보내기 위하여 수정버튼 handle함수를 용병 작성 컴포넌트로 이동
